+++ doc/changelog.md
@@ -26,6 +26,7 @@
 - Agent data for temperature and util are split into separate graphs and have more different colors.
 - Files can now be selected for either the cracker task or the preprocessor and are filled in the corresponding field.
 - Included new Hashcat modes included in newest beta.
+- Adjusted to new format of Hashcat printing cracked WPA hashes.
 
 # v0.10.1 -> v0.11.0